Finished first implementation of character saving
-Added some more cleaning functionality on the generated sentences
-Created temp loading script for loading saved characters into the sample scene
-Added system for adding an action prompt into the dialogue generator, to get a reaction in the generated text

+++ Assets/Scripts/AiAPI/GeneratorSettings.cs
@@ -16,15 +16,15 @@ public class GenerationSettings
     [HideInInspector][SerializeField] public bool use_world_info = false;
     [SerializeField] public int max_context_length = 1314;
     [SerializeField] public int max_length = 80;
-    [SerializeField] public float rep_pen = 1.06f;
+    [SerializeField] public double rep_pen = 1.06f;
     [SerializeField] public int rep_pen_range = 1600;
-    [SerializeField] public float rep_pen_slope = 0.9f;
-    [SerializeField] public float temperature = 0.69f;
-    [SerializeField] public float tfs = 1f; //Tail Free Sampling
-    [SerializeField] public float top_a = 0.0f; //Top A Sampling
-    [SerializeField] public float top_k = 0.0f; // Top K Sampling
-    [SerializeField] public float top_p = 0.69f; //Top P Sampling
-    [SerializeField] public float typical = 1.0f; //typical sampling
+    [SerializeField] public double rep_pen_slope = 0.9f;
+    [SerializeField] public double temperature = 0.69f;
+    [SerializeField] public double tfs = 1f; //Tail Free Sampling
+    [SerializeField] public double top_a = 0.0f; //Top A Sampling
+    [SerializeField] public double top_k = 0.0f; // Top K Sampling
+    [SerializeField] public double top_p = 0.69f; //Top P Sampling
+    [SerializeField] public double typical = 1.0f; //typical sampling
     [HideInInspector][SerializeField] public int[] sampling_order = { 6, 0, 1, 2, 3, 4, 5 };
     [SerializeField] public bool singleline = false;
